requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
pytest>=7.4.0
python-dateutil>=2.8.2
pyarrow>=12.0.0
